requirements.txt
PyYAML
pysigma
pySigma-backend-kusto
dotenv
openai
